requirements.txt
# Use: .\.venv\Scripts\python.exe -m pip install -r requirements.txt
streamlit>=1.30
numpy
opencv-python
tensorflow==2.20.0
joblib
scikit-learn==1.7.0
google-genai>=1.73
pandas
scipy
matplotlib
reportlab
